✨ Add all functions
