+++ src/components/Accodion.module.css
@@ -1,7 +1,7 @@
 .a {
   display: flex;
   flex-direction: column;
-  gap: 4px;
+  gap: 8px;
 }
 .accodion {
   min-height: 42px;
@@ -19,16 +19,13 @@
   height: 18px;
   display: flex;
   justify-content: space-between;
- 
   align-items: center;
   padding: 0;
- 
 }
 .text{
   font-weight: 400;
   font-size: 16px;
 }
-
 .accodionContainer {
   display: flex;
   flex-direction: column;
@@ -47,7 +44,6 @@
   border: 1px solid #E5E7EB;
   border-top-left-radius: 12px;
   border-top-right-radius: 12px;
-
 }
 .favorite {
   display: flex;
