+++ src/components/BestItem.jsx
@@ -1,20 +1,17 @@
-import styles from './AllItem.module.css';
-import AllItemList from './AllItemList';
+import styles from './BestItem.module.css';
+import BestItemList from './BestItemList';
 
-function BestItem({ items}) {
-
- 
+function BestItem({ items }) { 
   return (
     <div className={styles.section}>
       <h1 className={styles.titleText}>베스트 상품</h1>
       <ul className={styles.allItemBody}>
         {items.map((item) => (
           <li key={item.id} >
-            <AllItemList item={item} />
+            <BestItemList item={item} />
           </li>
         ))}
       </ul>
-
     </div>
   );
 }
