+++ src/components/Header.jsx
@@ -2,7 +2,6 @@ import logoImage from '../assets/logo.png';
 import profileImage from '../assets/profile.png';
 import styles from './Header.module.css';
 
-
 function Header() {
   return (
     <div className={styles.header}>
@@ -10,16 +9,12 @@ function Header() {
         <img className={styles.logo} src={logoImage} alt="" />
         <div className={styles.textBox}>
           <span className={styles.text}>자유게시판</span>
-          <span className={styles.text}>중고마켓</span>
+          <span className={`${styles.text} ${styles.blue}`}>중고마켓</span>
         </div>
-
       </div>
-   
       <img className={styles.profile} src={profileImage} alt="" />
     </div>
   );
-
-
 }
 
 export default Header;
