+++ src/components/Pagination.module.css
@@ -1,10 +1,13 @@
+.section {
+  max-height: 40px;
+  margin: 0 auto;
+  margin-bottom: 58px;
+}
 .paginationContainer {
   display: flex;
   justify-content: center;
   align-items: center;
-  margin: 30px 0;
 }
-
 .pageButton {
   border: 1px solid #ccc;
   background: white;
@@ -20,14 +23,12 @@
   align-items: center; 
   justify-content: center;
 }
-
 .active {
   background: #007bff;
   color: white;
   border-color: #007bff;
   font-weight: bold;
 }
-
 .pageButton:disabled {
   cursor: not-allowed;
   opacity: 0.5;
