+++ src/index.css
@@ -4,11 +4,14 @@
   font-weight: 400;
   color: #000;
   box-sizing: border-box;
+  margin: 0;
+  padding: 0;
 }
 
 body {
-    margin: 0;
-    padding: 0;
+    box-sizing: border-box;
+  margin: 0;
+  padding: 0;
 }
 
 /* index.css 또는 별도의 Font.css 파일 */
