+++ src/pages/AddItem.jsx
@@ -1,0 +1,9 @@
+function AddItem() {
+  return (
+    <>
+      <h1>AddItem Page</h1>
+    </>
+  );
+}
+
+export default AddItem;
